- Fixed code style according to code convention - Renamed AuthControl into RegistrationController - Add RegistrationService and TestClass for that - moved logic UserService and UserSecurityService to RegistrationService - Delete unnecessary UserSecurityService service

+++ src/main/java/com/ratifire/devrate/entity/Role.java
@@ -1,16 +1,19 @@
 package com.ratifire.devrate.entity;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
-
 /**
- * Entity class representing a role in the system.
- * A role defines a set of permissions or privileges granted to users.
+ * Entity class representing a role in the system. A role defines a set of permissions or privileges
+ * granted to users.
  */
 @Entity
 @Data
@@ -20,21 +23,21 @@ import java.math.BigInteger;
 @Table(name = "user_role")
 public class Role {
 
-    /**
-     * The unique identifier for the role.
-     */
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+  /**
+   * The unique identifier for the role.
+   */
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long id;
 
-    /**
-     * The name of the role.
-     */
-    @Column(nullable = false, unique = true)
-    private String name;
+  /**
+   * The name of the role.
+   */
+  @Column(nullable = false, unique = true)
+  private String name;
 
-    /**
-     * A brief description of the role.
-     */
-    private String description;
+  /**
+   * A brief description of the role.
+   */
+  private String description;
 }
